refactor(udp): refactor tests to follow mod structure conventions
Use the same dir/mod structure as in API and HTTP tracker integration
tests.
