+++ tests/common/mod.rs
@@ -1,2 +1,3 @@
 pub mod fixtures;
 pub mod http;
+pub mod udp;
